requirements.txt
contourpy==1.0.7
customplotting==0.1.4.dev0
cycler==0.11.0
fonttools==4.38.0
h5py==3.7.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
kiwisolver==1.4.4
libusb==1.0.26b5
libusb-package==1.0.26.2
matplotlib==3.6.3
matplotlib-scalebar==0.8.1
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pandas==2.1.4
Pillow==9.4.0
pkg-about==1.0.8
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
pyusb @ git+https://github.com/pyusb/pyusb@73e87f68fc8ece558057113690f9d2391441b58f
PyVISA==1.14.1
QtPy==2.3.0
ScopeFoundry==1.3.0
seabreeze==2.5.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
xreload==1.0.1
zipp==3.11.0
